requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.12.0
httpx==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
librosa>=0.10.0
soundfile>=0.12.0

